data-services/requirements.txt
fastf1==3.3.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.28
requests==2.31.0
beautifulsoup4==4.12.3
schedule==1.2.1
python-dotenv==1.0.1
redis==5.0.3
